requirements.txt
datetime 
PyMySQL
numpy
mysql-connector-python
-e .

